Add vimrc and theme, update install script
The vim dir will is designed to be symlinked to .vim in ~/. The
install script will now check if any of the files already exist
in ~/, and if they do it will remove them and symlink the files
in the current working directory. If the config files do not exist
in ~/, it will create the symlinks.